fix bug(积分、优惠券、购物券类型改为BigDecimal)

+++ distribution-parent/distribution-data/src/main/java/com/chauncy/data/vo/manage/activity/gift/FindGiftVo.java
@@ -35,13 +35,13 @@ public class FindGiftVo {
     private String name;
 
     @ApiModelProperty(value = "经验值")
-    private Integer experience;
+    private BigDecimal experience;
 
     @ApiModelProperty(value = "购物券")
-    private Integer vouchers;
+    private BigDecimal vouchers;
 
     @ApiModelProperty(value = "积分")
-    private Integer integrals;
+    private BigDecimal integrals;
 
     @ApiModelProperty(value = "购买金额,新人礼包不显示")
     private BigDecimal purchasePrice;
